Use uint instead of size_t to store token and AST positions in C++ parser and the duchain. This saves memory on x86_64 systems because size_t is twice as big as uint on Unix/LP64.
Optimization results:
peak memory consumption for parsing Ruby interpreter sources:
    before: 909M
    after: 808M
101M memory saved

+++ languages/cpp/parser/commentformatter.cpp
@@ -25,18 +25,18 @@
 #include <language/duchain/stringhelpers.h>
 
 
-QByteArray CommentFormatter::formatComment( size_t token, const ParseSession* session ) {
+QByteArray CommentFormatter::formatComment( uint token, const ParseSession* session ) {
   if( !token )
     return QByteArray();
   const Token& commentToken( (*session->token_stream)[token] );
   return KDevelop::formatComment( stringFromContents(session->contentsVector(), commentToken.position, commentToken.size ) );
 }
 
-QByteArray CommentFormatter::formatComment( const ListNode<size_t>* comments, const ParseSession* session ) {
+QByteArray CommentFormatter::formatComment( const ListNode<uint>* comments, const ParseSession* session ) {
   QByteArray ret;
   if( comments )
   {
-    const ListNode<size_t> *it = comments->toFront(), *end = it;
+    const ListNode<uint> *it = comments->toFront(), *end = it;
     do {
       QByteArray c = CommentFormatter::formatComment(it->element, session);
